- spent some time fixing the marginal demand and sell math in EconomyMath and Market and Mayor - moved addTradedResource to its own method and called it from both constructors - fixed deserialization mayor issues - moved mayor spawn egg to command, removed egg. - removed adopt mayor code, just adds village where command is called

+++ common/src/main/java/com/holybuckets/villageecon/core/model/ResourceLedger.java
@@ -91,7 +91,7 @@ public class ResourceLedger {
      */
     public void logTrade(Sale sale, boolean isBuyer) {
         if (sale == null) return;
-        int totalPrice = sale.getSalePrice() * sale.getSaleQuantity();
+        float totalPrice = sale.getSalePrice() * sale.getSaleQuantity();
         if (isBuyer) {
             add(sale.getResourceId(), sale.getSaleQuantity());
             addCurrency(-totalPrice);
